replace appdb.py with fedcloud_is.py
